refactor: alterando os parâmentos width, height, mass e speed de float para int

+++ src/com/gamestudio/elements/Bat.java
@@ -11,7 +11,7 @@ public class Bat extends DumbRobot {
     private Animation idleAnim;
     private Animation openingWingsAnim;
     private Animation flyingAnim;
-    private float speed;
+    private int speed;
     private boolean isActive;
     private boolean isWingsOpened;
